refactor: Move event listener into LootrunModel
